refactor TimeBoundaryInfo to pinot-core
